Split custom attributes code to separate files and use embedded struct for shared attributes
